done updating .net core 2.1 to 3.1 - added Data Lookup successfully
