+++ BillingNextSys/BillingNextSys/BillingNextSys.csproj
@@ -6,19 +6,23 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Datalist" Version="7.2.1" />
     <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="3.1.1" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.1" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="3.1.1" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.1" />
+    <PackageReference Include="Microsoft.AspNetCore.Session" Version="2.2.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="3.1.1" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.0" PrivateAssets="All" />
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.1" PrivateAssets="All" />
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.4" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.1.0" />
+    <PackageReference Include="NonFactors.Lookup.Mvc6" Version="4.2.1" />
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.1.1" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL.Design" Version="1.1.1" />
     <PackageReference Include="NonFactors.Grid.Mvc6" Version="5.1.2" />
-    <PackageReference Include="Twilio" Version="5.25.1" />
+    <PackageReference Include="Twilio" Version="5.37.3" />
     <PackageReference Include="EPPlus.Core" Version="1.5.4" />
-    <PackageReference Include="RestClient" Version="2.4.4.21" />
-    <PackageReference Include="RestRequest" Version="1.1.2" />
+    <PackageReference Include="RestClient" Version="2.5.0.6" />
+    <PackageReference Include="RestRequest" Version="1.1.4" />
     <PackageReference Include="IEnterprise.Easy-Http" Version="1.1.1" />
   </ItemGroup>
 <ItemGroup>
@@ -36,7 +40,6 @@
   <Folder Include="Pages\Particulars\" />
   <Folder Include="Pages\Received\" />
   <Folder Include="Pages\BillSeries\" />
-  <Folder Include="wwwroot\Content\" />
   <Folder Include="Pages\Bill\Format1\" />
   <Folder Include="Pages\Bill\Format2\" />
   <Folder Include="Pages\Export\" />
