+++ BillingNextSys/BillingNextSys/Pages/Bill/Format1/Create.cshtml.cs
@@ -128,12 +128,12 @@ namespace BillingNextSys.Pages.Bill.Format1
         {
             try
             {
-                obj.CompanyID = (int)_session.GetInt32("Cid");
-                obj.BranchID = (int)_session.GetInt32("Bid");
                 Random generator = new Random();
                 obj.SecretUnlockCode =Int32.Parse(generator.Next(0, 999999).ToString("D6"));
                 obj.BillDelivered = false;
                 obj.MessageSent = false;
+                obj.CompanyID = (int)_session.GetInt32("Cid");
+                obj.BranchID = (int)_session.GetInt32("Bid");
                 obj.BillDate = DateTime.Now;
                 _context.Bill.Add(obj);
                 _context.SaveChanges();
