+++ BillingNextSys/BillingNextSys/Pages/Shared/MvcLookup/_Dialog.cshtml
@@ -1,0 +1,33 @@
+﻿<div class="mvc-lookup-overlay">
+    <div class="mvc-lookup-wrapper">
+        <div id="MvcLookupDialog" class="mvc-lookup-dialog">
+            <div class="mvc-lookup-header">
+                <span class="mvc-lookup-title"></span>
+                <button class="mvc-lookup-close">×</button>
+            </div>
+            <div class="mvc-lookup-body">
+                <div class="mvc-lookup-search-area">
+                    <input class="mvc-lookup-search" />
+                    <input class="mvc-lookup-rows" />
+                    <button class="mvc-lookup-selector"></button>
+                </div>
+                <div class="mvc-lookup-content">
+                    <div class="mvc-lookup-dialog-loader">
+                        <span></span>
+                    </div>
+                    <div class="mvc-lookup-dialog-error">
+                    </div>
+                    <div class="mvc-lookup-table">
+                        <table>
+                            <thead></thead>
+                            <tbody></tbody>
+                        </table>
+                    </div>
+                    <div class="mvc-lookup-footer">
+                        <button type="button" class="mvc-lookup-load-more"></button>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
