+++ BillingNextSys/BillingNextSys/Pages/Shared/_LayoutPrint.cshtml
@@ -16,7 +16,7 @@
               asp-fallback-test-class="sr-only" asp-fallback-test-property="position" asp-fallback-test-value="absolute" />
         <link rel="stylesheet" href="~/css/site.min.css" asp-append-version="true" />
     </environment>
-    <link href="~/Content/MvcGrid/mvc-grid.css" rel="stylesheet">
+    <link href="~/Content/mvc-grid/mvc-grid.css" rel="stylesheet">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 
 </head>
@@ -39,7 +39,7 @@
         @RenderBody()
         <hr />
         <footer>
-            <p class="pull-right">&copy; 2019 - BillingNext - Developed By Ruchit Patel</p>
+            <p class="pull-right">&copy; 2020 - BillingNext - Developed By Ruchit Patel</p>
         </footer>
     </div>
     <environment include="Development">
