+++ BillingNextSys/BillingNextSys/Pages/_ViewImports.cshtml
@@ -1,4 +1,6 @@
 @using BillingNextSys
 @using NonFactors.Mvc.Grid
+@using NonFactors.Mvc.Lookup;
 @namespace BillingNextSys.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@addTagHelper *, Mvc.Lookup.Core // Optional for tag helpers
